requirements.txt
altair==5.5.0
atlassian-python-api==4.0.4
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
boto3==1.38.36
botocore==1.38.36
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.4
Deprecated==1.2.18
detect-secrets==1.5.0
elastic-transport==8.17.1
elasticsearch==9.0.2
Flask==3.1.1
gitdb==4.0.12
gitdb2==4.0.2
GitPython==3.1.44
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
narwhals==1.42.1
numpy==2.3.0
oauthlib==3.2.2
packaging==24.2
pandas==2.3.0
pillow==11.2.1
pluggy==1.6.0
protobuf==6.31.1
pyarrow==20.0.0
pycparser==2.22
pydeck==0.9.1
PyGithub==2.6.1
Pygments==2.19.1
PyJWT==2.10.1
PyNaCl==1.5.0
pytest==8.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.0.0
rpds-py==0.25.1
s3transfer==0.13.0
scikit-learn==1.7.0
scipy==1.15.3
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
soupsieve==2.7
streamlit==1.45.1
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.1
truffleHog==2.2.1
truffleHogRegexes==0.0.7
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
